fixed hanging reverse shell on windows
When creating a shell using exec.Command(), we opened a stderr pipe for the process but never used it or read from it. This didn't come up as a problem on UNIX, but on Windows the pipe only has a buffer of 4096 bytes. As a result, it would fill up after a while and block any more output from being written to the STDOUT pipe, causing the shell to mysteriously "hang" after a while. Removing stderr from command.go was enough to fix this.