fixing shell input buffers
When reading input, we were making a 500-byte buffer for commands and not checking how many bytes were actually written to it.
As a result, there were lots of trailing null bytes that broke the strings.TrimRight() call meant to strip away newlines from the end of the string.
This meant that two newlines were being submitted for each command.